finished refactoring select dataslot
