+++ public/views/SelectTraining.vue
@@ -97,6 +97,12 @@
             :data-items="items"
             :data-fields="fields"
             :timeseries="timeseries"
+            :baseline-items="baselineItems"
+            :inner="inner"
+            :outer="outer"
+            @fetch-variable-rankings="fetchVariableRankings"
+            @tile-clicked="onTileClick"
+            @fetch-timeseries="fetchTimeseries"
           />
           <create-solutions-form class="select-create-solutions" />
         </div>
@@ -120,6 +126,8 @@ import {
   TimeSeries,
   TableRow,
   TableColumn,
+  D3M_INDEX_FIELD,
+  TimeseriesGrouping,
 } from "../store/dataset/index";
 import TargetVariable from "../components/TargetVariable.vue";
 import { DataMode } from "../store/dataset";
@@ -136,8 +144,10 @@ import {
   TRAINING_VARS_INSTANCE,
   AVAILABLE_TRAINING_VARS_INSTANCE,
 } from "../store/route/index";
+import { Filter, EXCLUDE_FILTER } from "../util/filters";
 import { Dictionary } from "../util/dict";
 import { Route } from "vue-router";
+import { EI } from "../util/events";
 import { Group } from "../util/facets";
 
 export default Vue.extend({
@@ -263,6 +273,23 @@ export default Vue.extend({
         this.availableTrainingVarsSearch
       );
     },
+    baselineItems(): TableRow[] {
+      const bItems =
+        datasetGetters.getBaselineIncludeTableDataItems(this.$store) ?? [];
+      return bItems.sort((a, b) => {
+        return a[D3M_INDEX_FIELD] - b[D3M_INDEX_FIELD];
+      });
+    },
+    inner(): TableRow[] {
+      return this.includedActive
+        ? datasetGetters.getAreaOfInterestIncludeInnerItems(this.$store)
+        : datasetGetters.getAreaOfInterestExcludeInnerItems(this.$store);
+    },
+    outer(): TableRow[] {
+      return this.includedActive
+        ? datasetGetters.getAreaOfInterestIncludeOuterItems(this.$store)
+        : datasetGetters.getAreaOfInterestExcludeOuterItems(this.$store);
+    },
     variableDict(): Dictionary<Dictionary<VariableSummary>> {
       return this.includedActive
         ? datasetGetters.getIncludedVariableSummariesDictionary(this.$store)
@@ -285,8 +312,14 @@ export default Vue.extend({
 
       return currentSummaries;
     },
+    dataSize(): number {
+      return routeGetters.getRouteDataSize(this.$store);
+    },
   },
   watch: {
+    dataSize() {
+      viewActions.updateSelectTrainingData(this.$store);
+    },
     trainingStr() {
       viewActions.updateSelectTrainingData(this.$store);
     },
@@ -332,6 +365,43 @@ export default Vue.extend({
     viewActions.updateHighlight(this.$store);
   },
   methods: {
+    fetchTimeseries(args: EI.TIMESERIES.FetchTimeseriesEvent) {
+      args.variables.forEach((tsv) => {
+        const grouping = tsv.grouping as TimeseriesGrouping;
+        datasetActions.fetchTimeseries(this.$store, {
+          dataset: this.dataset,
+          variableKey: tsv.key,
+          xColName: grouping.xCol,
+          yColName: grouping.yCol,
+          uniqueTrail: args.uniqueTrail,
+          timeseriesIds: args.timeseriesIds.map((item) => {
+            return item[tsv.key].value as string;
+          }),
+        });
+      });
+    },
+    async onTileClick(data: EI.MAP.TileClickData) {
+      // filter for area of interests
+      const filter: Filter = {
+        displayName: data.displayName,
+        key: data.key,
+        maxX: data.bounds[1][1],
+        maxY: data.bounds[0][0],
+        minX: data.bounds[0][1],
+        minY: data.bounds[1][0],
+        mode: EXCLUDE_FILTER,
+        type: data.type,
+        set: "",
+      };
+      // fetch area of interests
+      await viewActions.updateAreaOfInterest(this.$store, filter);
+    },
+    fetchVariableRankings() {
+      datasetActions.fetchVariableRankings(this.$store, {
+        dataset: this.dataset,
+        target: this.target,
+      });
+    },
     async addVar(group: Group) {
       this.$router;
       addVariableToTraining(group, this.$router);
